+++ src/geometry/mod.rs
@@ -27,8 +27,8 @@ use rapier::prelude::Group;
 
 pub const fn unpack_interaction_groups(memberships_filter: u32) -> InteractionGroups {
     InteractionGroups::new(
-        unsafe { Group::from_bits_unchecked((memberships_filter >> 16) as u32) },
-        unsafe { Group::from_bits_unchecked((memberships_filter & 0x0000_ffff) as u32) },
+        Group::from_bits_retain((memberships_filter >> 16) as u32),
+        Group::from_bits_retain((memberships_filter & 0x0000_ffff) as u32),
     )
 }